Refactor: create CategoriesContainer.jsx
