+++ src/App.jsx
@@ -1,17 +1,7 @@
-export default function App() {
-  const categories = [
-    { id: 1, name: '한식' },
-  ];
+import CategoriesContainer from './CategoriesContainer';
 
+export default function App() {
   return (
-    <div>
-      <ul>
-        {categories.map((category) => (
-          <li key={category.id}>
-            {category.name}
-          </li>
-        ))}
-      </ul>
-    </div>
+    <CategoriesContainer />
   );
 }
